Refactor : Enter Room Logic and WebSocketConnection code Refactoring
